// src/providers/generated/usps.ts
export interface components {
    schemas: {
        /** @description An array of detailed tracking event information for the requested tracking number in reverse chronological order. */
        MultiStatusResponse: (components["schemas"]["FailureResponse"] | components["schemas"]["SuccessResponse"])[];
// ...
        TrackingDetails: components["schemas"]["TrackingDetail"][];
        /** Tracking Request Parameters */
        TrackingRequest: components["schemas"]["TrackingRequestBody"][];
// ...
        ErrorMessage: {
            /** @description The version of the API that was used and that raised the error. */
            apiVersion?: string;
            /** @description The high-level error that has occurred as indicated by the status code. */
            error?: {
                /** @description The error status code that has been returned in response to the request. */
                code?: string;
                /** @description A human-readable message describing the error. */
                message?: string;
                errors?: ({
                    /** @description The status code response returned to the client. */
                    status?: string;
                    /** @description An internal subordinate code used for error diagnosis. */
                    code?: string;
                    /** @description A human-readable title that identifies the error. */
                    title?: string;
                    /** @description A human-readable description of the error that occurred. */
                    detail?: string;
                    /** @description The element that is suspected of originating the error.  Helps to pinpoint the problem. */
                    source?: {
                        /** @description The input in the request which caused an error. */
                        parameter?: string;
                        /** @description An example of a valid value for the input parameter. */
                        example?: string;
                    } & {
                        [key: string]: unknown;
                    };
                } & {
                    [key: string]: unknown;
                })[];
            } & {
                [key: string]: unknown;
            };
        } & {
            [key: string]: unknown;
        };
// ...
        TrackingDetail: {
            [key: string]: unknown;
        } & (components["schemas"]["SharedTrackingDetail"] & {
            deliveryDateExpectation?: components["schemas"]["DeliveryDateExpectation"];
        } & {
            servicesEligibility?: components["schemas"]["ServicesEligibility"];
        } & {
            trackingEvents?: components["schemas"]["TrackingEvents"];
        });
// ...
        TrackingEvent: {
            /** @description Action code for the event. */
            actionCode?: string;
            /** @description Collect on delivery amount due. Only applicable for Arrival at Unit (07) or Available for Pickup (14) events. */
            CODAmountDue?: string;
            /** @description The city where the event occurred. */
            eventCity?: string;
            /** @description Tracking event code. Refer to Domestic Scan Event list provided in USPS&#174; [Pub 199](http://postalpro.usps.com/pub199) Appendix G for all possible values. */
            eventCode?: string;
            /** @description The country where the event occurred. */
            eventCountry?: string;
            /** @description The name of the partner associated to the event. */
            eventPartner?: string;
            /** @description The state where the event occurred. */
            eventState?: string;
            eventTimestamp?: components["schemas"]["AllEventDate"];
            /** @description The description of the event (e.g. Delivery). */
            eventType?: string;
            /**
             * @description The ZIP Code&#8482; of the event.
             * @example 12345
             */
            eventZIPCode?: string;
            /** @description The company name if delivered to a company. */
            firm?: string;
            /**
             * Format: date-time
             * @description The date and time of the event in Greenwich Mean Time (GMT).
             * @example 2024-04-04T14:03:12.041Z
             */
            GMTTimestamp?: string;
            /**
             * @description The offset of the local time of the event from Greenwich Mean Time (GMT).
             * @example -7:00
             */
            GMTOffset?: string;
            /** @description Reason code for the event. */
            reasonCode?: string;
            /** @description The name of the person signing for delivery (if available). */
            recipientName?: string;
            /** @description Delivery scan location certified, if applicable. */
            veriPoint?: boolean;
        };
// ...
    };
// ...
}